% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kmerDistance.dif}
\alias{kmerDistance.dif}
\title{calculating pairwise distance between genomes base on kmer stratege}
\usage{
kmerDistance.dif(kmer.length, path.to.data = "../data")
}
\arguments{
\item{kmer.length}{length of kmer.}

\item{path.to.data}{path to test data which are integer kc files generated by using Kanalyze. Default to '../data/' folder.}
}
\value{
a matrix
}
\description{
Returns a pairwise distance matrix.
}
\details{
This function compute pairwise distnace of kmer data by using XOR discrete funtion.
}

