% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kmerDistance.hpca}
\alias{kmerDistance.hpca}
\title{calculating pairwise distance between genome base on kmer stratege}
\usage{
kmerDistance.hpca(kmer.length, path.to.data = "../data", ifscale = TRUE,
  ifbinary = FALSE, numofp = 3, ifplot = FALSE, plotmain = "Main")
}
\arguments{
\item{kmer.length}{length of kmer.}

\item{path.to.data}{path to test data. Test data are integer kc files generated by using Kanalyze. Default to '../data/' folder}

\item{ifscale}{scale flag. if this flag is set to true, the data is scaled.}

\item{ifbinary}{binary flag. if this flag is set true, the kmer counts are set to binary number.}

\item{numofp}{number of components used to calculate the distance.}

\item{ifplot}{plot flag. if this flag is set to true, and numofp is equal to 3, a 3d plot will be generate base on the main components.}

\item{plotmain}{The main title (on top) using font and size (character expansion) par("font.main") and color par("col.main").}
}
\value{
a matrix
}
\description{
Returns a pairwise distance matrix.
}
\details{
This function compute pairwise distnace of kmer data by using hierarchical PCA.
}

